# PERF: Event emissions and perf regression.