@Czaki 2022-11-06
- but the comparison is of two instances of `TextManager`, not empty dicts.