@Carreau 2022-11-06
> * but the comparison is of two instances of `TextManager`, not empty dicts.

Sure, but I still came across this while it was comparing empty dicts due to register_theme.
And optimising that had a non negligible impact to compare TextManager instances. And even with TextManager instances, I'm not sure why it does skip identity check...

This is also why I added the profile of `TextManager.__init__`, to show that it improves that.