@Czaki 2022-11-06
> Sure, but I still came across this while it was comparing empty dicts due to register_theme.

I do not understand. Where is a problem with comparing empty dicts? I do not see it in the profiling output.  


> And optimising that had a non negligible impact to compare TextManager instances. And even with TextManager instances, I'm not sure why it does skip identity check...

Because it is how `__eq__` operator is implemented: 

https://github.com/napari/napari/blob/1cc90a74211c6cbf3fc644da5866f9752cf1ac78/napari/utils/events/evented_model.py#L325-L345

I never expected that python would use `is` operator before calling `__eq__` operator. But I may be wrong.