@Carreau 2022-11-06
> Because it is how `__eq__` operator is implemented:

I'm (was) pretty sure that is an optimisation at CPython level for `==`, before even calling `__eq__`, or maybe I'm confusing with PyPy, or another thing like set belonging.