@Carreau 2022-11-06
Ah, Yes, there is a shortcut for `x in [a list with x]`, according to https://stackoverflow.com/questions/49429308/should-i-test-for-object-identity-in-eq-function