@brisvag 2022-11-07
Should we then simply add the identity check at the top of our `__eq__` implementation, to avoid this issue in similar cases?

Yeah it seems like a useful optimization to have... On the other hand, this would break stuff like:
```py
np.nan != np.nan
```