@Carreau 2022-11-07
> Should we then simply add the identity check at the top of our `__eq__` implementation

Problem is that there are many sources, and you would need to add the identity check everywhere. In the case of `nan`, it's a problem, but I'm unsure `nan` can emit events. The other questions is do we _actually_ want to check for equality here, or are we actually looking for identity ? It looks to me that really we are checking that the last remaining source `is` the object emitting the event, so it might be that `is` is actually correct while `==` is not...