@brisvag 2022-11-07
> Problem is that there are many sources, and you would need to add the identity check everywhere.

Not if the problem is the slowness of `EventedModel`'s eq. You can just add it at the top of that `eq`, and at least you cut down that cost, and you get the benefit in every other place where we might do `model == model`.

> In the case of `nan`, it's a problem, but I'm unsure `nan` can emit events.

Sorry, I wasn't clear. I meant that if the interpreter used this shortcut by default, it would not allow things like `nan` where `is` returns `True`, but `==` returns `False`.

> The other questions is do we _actually_ want to check for equality here, or are we actually looking for identity ? It looks to me that really we are checking that the last remaining source `is` the object emitting the event, so it might be that `is` is actually correct while `==` is not...

Actually, I think you're right. Either way, I think the addition to `__eq__` I suggested above would still be useful for other cases.