@Carreau 2022-11-09
Ok, I've move the instance test to the EventedModel `__eq__`, I didn't touch non-evented model classes, maybe some of them could use the same instance testing but should be reviewed on a case by case basis. 

I've also asserted that sources that don't test equal still are instances of EventedModel