@Carreau 2022-12-21
> you have an approval on this, just bringing it to attention in case you forgot about it :)

I try to always avoid self-merge, and try to follow the policy that only @napari/core-devs should merged – at least I was reminded some time ago I should not merge, so I'm not merging.